requirements.txt
annotated-types==0.5.0
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
peewee==3.16.2
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.3
